Re-await on failed await.
The problem with having a continuation for the `Throw` constructor is
how to expose the functionality to the user.

With this commit, every await becomes an await-throw loop when it fails.
This makes the natural definition of `pipe`, `idP` and `forP` do the
right thing, i.e. act as proper "filters" when composed on the left.