Add upstream return type as an extra parameter for Pipe.
First stab at implementing Chris Smith's idea of a fifth type parameter
represeting the upstream return value.

Unlike Chris's sketch, this implementation tries to provide a simple
`await` primitive:

    await :: Pipe a b u m a

which uses the upstream return value when it fails because of downstream
termination.

Unfortunately, this constrains the type of composition to

    (>+>) :: Pipe a b r m r -> Pipe b c r m s -> Pipe a c r m s

and makes a number of combinator impossible to use (see tests for
examples).