Fix broken finally implementation.
There were two bugs in `finally`.

First, `finallyP` was not recursing over the continuation of `Yield` and
`Throw` constructors, so finalizers didn't survive after a `yield`.

Second, `finally` was calling the finalizer after the end of the
`finallyP` block. But `finallyP` attaches the finalizer to the last
`Pure` constructor, so that was obviously incorrect.

Also add a regression test.